integration/requirements.txt
cattle==0.5.1
requests
pycrypto==2.6.1

flake8
pytest==2.3.5
